Created BlackJackCardHand and ICardHand which it inherits from
Handels the "card hands" in the game. both player and NPC cardhands.
this includes the calculating of the hand value, which takes into account that knights, queens, and kings all have the value of 10, and ace can have a value of either 1 or 11, depending on the current hands total value.