base for testing has been implemented.
The Model file GameLogic will act as a mediator between the viewModels and the rest of the model files.
It creates the card deck and hands that are to be used.
the bare minimum of what is needed to se a card and the buttons have been implemented in MainWindow.

i will continue to add functionality allowing more cards to be added and new views when more hands are added.