Updated the IPlayerHands and PlayerHands, bool is now returned
Updated the IPlayerHands and PlayerHands to now return a bool  if the split was successful.
The split logic was also simplified as the initial version was constructed after a misunderstanding of the rules.
Knowing that a split can only be conducted with the first two cards meant that we already know which cards is to be split and removed.