Changes to ICard, CardHandViewModel , CardHandView
I now send the whole cardHand object to the viewmodel, instead of only the list. Allowing access to handvalues.
I also included the bool variable FaceDown to the ICard Interface, as it is needed when performing hand value calculations.
When the card if face down, it should not be counted until it is face up.