Moved Player interaction logic from gameLogic to a new class PlayerRound
On a new round gameLogic sends the chosen playerHand to PlayerRound.StartNewRound which handles the player actions asynchronously.

In case of a split, this class is notified through MessageBuss<SplitSuccessfulMessage>, where the new hand is added to the queue of hands to process.

unlike how it was built before, now a public Subject<playeraction> is used to this class when the player has chosen an action (buttonViewModel)