Added A Player class, major rework of the code to handle more then one player has been conducted.
The Player class contain the attributes
Name
The players CardHands
Funds
and methods for regestering bet and payout.

The split functionality is not in a functional state because of the change. I plan to implement a PlayerViewModel which will be used to present the cardviews related to a Player. It is also there that the split will be handled in regards to viewmodels.